requirements_dev.txt
flake8
mock
coverage
django-coverage
transifex-client
coveralls
sphinx
tox
django-nose
.
